+++ apps/web/src/main.tsx
@@ -2,9 +2,8 @@ import './index.css'
 
 import React from 'react'
 import ReactDOM from 'react-dom/client'
-import { VinessReactApp } from '@viness/core'
+import { AppRouter, VinessReactApp } from '@viness/core'
 import { app } from './app'
-import { router } from './routes'
 
 const ErrorFallbackComponent = () => <div>error</div>
 const SuspenseFallbackComponent = () => <div></div>
@@ -17,8 +16,7 @@ ReactDOM.createRoot(document.getElementById('root') as HTMLElement).render(
             ErrorFallbackComponent={ErrorFallbackComponent}
         >
             {/* Other Custom Providers Wrap Routes */}
-
-            {router}
+            <AppRouter />
         </VinessReactApp>
     </React.StrictMode>
 )