+++ packages/viness-core/src/app-react-hooks.tsx
@@ -1,6 +1,7 @@
 import { ServiceIdentifier } from '@viness/di'
 import { VinessApp } from './app'
 import { useAppContext } from './app-react-context'
+import { VinessRoute } from './route'
 
 /**
  * Get the Viness App instance
@@ -17,4 +18,9 @@ export function useService<T>(id: ServiceIdentifier<T>) {
     const app = useAppContext()
     const container = app?.container
     return container?.get(id) as T
+}
+
+export function useRoute(id: ServiceIdentifier<VinessRoute>) {
+    const app = useAppContext()
+    return app?.getService(id)
 }