+++ packages/viness-core/src/index.tsx
@@ -1,6 +1,8 @@
 export * from './app'
 export * from './app-react'
 export * from './app-react-hooks'
-export * from './route'
+export * from './router-react'
 export * from './container'
 export * from './ui-store'
+
+export { useNavigate, useLocation, useParams, Outlet } from 'react-router-dom'